+++ Tests/IEnumerableTests.cs
@@ -5,9 +5,9 @@ public class EnumerableTests {
 
   [Test]
   public void Enumerable_FiniteCFraction_CorrectCoefficients() {
-    CFraction fraction       = CFraction.FromRational(10, 7); // [1; 2, 3]
-    List<int> expectedCoeffs = new List<int> { 1, 2, 3 };
-    List<int> actualCoeffs   = new List<int>();
+    CFraction        fraction       = CFraction.FromRational(10, 7); // [1; 2, 3]
+    List<BigInteger> expectedCoeffs = new List<BigInteger> { 1, 2, 3 };
+    List<BigInteger> actualCoeffs   = new List<BigInteger>();
 
     foreach (int coeff in fraction) {
       actualCoeffs.Add(coeff);
@@ -18,8 +18,8 @@ public class EnumerableTests {
 
   [Test]
   public void Enumerable_EmptyCFraction_NoCoefficients() {
-    CFraction fraction     = CFraction.FromCoeffs(new List<int>()); // Empty CFraction
-    List<int> actualCoeffs = new List<int>();
+    CFraction        fraction     = CFraction.FromCoeffs(new List<BigInteger>()); // Empty CFraction
+    List<BigInteger> actualCoeffs = new List<BigInteger>();
 
     foreach (int coeff in fraction) {
       actualCoeffs.Add(coeff);
@@ -31,8 +31,8 @@ public class EnumerableTests {
   [Test]
   public void Enumerable_ResetEnumerator_IterateFromBeginning() {
     CFraction fraction = CFraction.FromRational(10, 7); // [1; 2, 3]
-    List<int> expectedCoeffs =
-      new List<int>
+    List<BigInteger> expectedCoeffs =
+      new List<BigInteger>
         {
           1
         , 2
@@ -41,7 +41,7 @@ public class EnumerableTests {
         , 2
         , 3
         };
-    List<int> actualCoeffs = new List<int>();
+    List<BigInteger> actualCoeffs = new List<BigInteger>();
 
     using var enumerator = fraction.GetEnumerator();
 
@@ -60,21 +60,21 @@ public class EnumerableTests {
 
   [Test]
   public void Enumerable_ManualMoveNext_CurrentValuesCorrect() {
-    CFraction fraction       = CFraction.FromRational(10, 7); // [1; 2, 3]
-    List<int> expectedCoeffs = new List<int> { 1, 2, 3 };
-    List<int> actualCoeffs   = new List<int>();
-    var       enumerator     = fraction.GetEnumerator();
+    CFraction        fraction       = CFraction.FromRational(10, 7); // [1; 2, 3]
+    List<BigInteger> expectedCoeffs = new List<BigInteger> { 1, 2, 3 };
+    List<BigInteger> actualCoeffs   = new List<BigInteger>();
+    var              enumerator     = fraction.GetEnumerator();
 
     Assert.That(enumerator.MoveNext(), Is.True, "MoveNext should return true for the first coefficient");
-    Assert.That(enumerator.Current, Is.EqualTo(1), "Current should be the first coefficient");
+    Assert.That(enumerator.Current, Is.EqualTo((BigInteger)1), "Current should be the first coefficient");
     actualCoeffs.Add(enumerator.Current);
 
     Assert.That(enumerator.MoveNext(), Is.True, "MoveNext should return true for the second coefficient");
-    Assert.That(enumerator.Current, Is.EqualTo(2), "Current should be the second coefficient");
+    Assert.That(enumerator.Current, Is.EqualTo((BigInteger)2), "Current should be the second coefficient");
     actualCoeffs.Add(enumerator.Current);
 
     Assert.That(enumerator.MoveNext(), Is.True, "MoveNext should return true for the third coefficient");
-    Assert.That(enumerator.Current, Is.EqualTo(3), "Current should be the third coefficient");
+    Assert.That(enumerator.Current, Is.EqualTo((BigInteger)3), "Current should be the third coefficient");
     actualCoeffs.Add(enumerator.Current);
 
     Assert.That(enumerator.MoveNext(), Is.False, "MoveNext should return false after the last coefficient");
@@ -85,32 +85,32 @@ public class EnumerableTests {
 #region Cache and Enumerable Interaction Tests
   [Test]
   public void Enumerable_PartialCache_FullIteration() {
-    CFraction fraction      = CFraction.E;
-    var       _             = fraction[1];
-    List<int> actualCoeffs  = new List<int>();
-    int       expectedCount = 10;
+    CFraction        fraction      = CFraction.E;
+    var              _             = fraction[1];
+    List<BigInteger> actualCoeffs  = new List<BigInteger>();
+    int              expectedCount = 10;
 
     foreach (int coeff in fraction.Take(expectedCount)) {
       actualCoeffs.Add(coeff);
     }
 
-    List<int> expectedCoeffs = CFraction.EGenerator().Take(expectedCount).ToList();
+    List<BigInteger> expectedCoeffs = CFraction.EGenerator().Take(expectedCount).ToList();
     Assert.That
       (actualCoeffs, Is.EqualTo(expectedCoeffs), "Full iteration after partial cache should return correct coefficients");
   }
 
   [Test]
   public void Enumerable_CacheExpansionDuringIteration() {
-    CFraction fraction       = CFraction.FromRational(142, 43); // [3; 3, 3, 4]
-    List<int> actualCoeffs   = new List<int>();
-    List<int> expectedCoeffs = new List<int> { 3, 3, 3, 4 };
+    CFraction        fraction       = CFraction.FromRational(142, 43); // [3; 3, 3, 4]
+    List<BigInteger> actualCoeffs   = new List<BigInteger>();
+    List<BigInteger> expectedCoeffs = new List<BigInteger> { 3, 3, 3, 4 };
 
     foreach (int coeff in fraction) {
       actualCoeffs.Add(coeff);
     }
 
     Assert.That(actualCoeffs, Is.EqualTo(expectedCoeffs), "Iteration should expand cache and return all coefficients");
-    Assert.That(fraction[3], Is.EqualTo(4), "Cache should be fully populated after iteration");
+    Assert.That(fraction[3], Is.EqualTo((BigInteger)4), "Cache should be fully populated after iteration");
   }
 
   [Test]
@@ -120,11 +120,11 @@ public class EnumerableTests {
     _ = fraction[0];
     _ = fraction[1];
 
-    List<int> actualCoeffs = new List<int>();
+    List<BigInteger> actualCoeffs = new List<BigInteger>();
     foreach (int coeff in fraction) {
       actualCoeffs.Add(coeff);
     }
-    List<int> expectedCoeffs = new List<int> { 3, 7, 16 };
+    List<BigInteger> expectedCoeffs = new List<BigInteger> { 3, 7, 16 };
     Assert.That(actualCoeffs, Is.EqualTo(expectedCoeffs), "Iteration after indexer access should return correct coefficients");
   }
 
@@ -133,8 +133,8 @@ public class EnumerableTests {
     CFraction fraction = CFraction.E;
     _ = fraction.Take(5);
 
-    List<int> nextCoeffs         = fraction.Skip(5).Take(5).ToList();
-    List<int> expectedNextCoeffs = CFraction.EGenerator().Take(10).Skip(5).ToList();
+    List<BigInteger> nextCoeffs         = fraction.Skip(5).Take(5).ToList();
+    List<BigInteger> expectedNextCoeffs = CFraction.EGenerator().Take(10).Skip(5).ToList();
     Assert.That
       (
        nextCoeffs
